学生修改详情模块

+++ src/main/java/com/hudongwx/studentsys/controller/StudentController.java
@@ -110,7 +110,7 @@ public class StudentController extends BaseController {
             JSONObject jsonObject = JSON.parseObject(o.toString());
             Student stu = new Student();
             stu.setId(jsonObject.getInteger("id"));
-            stu.setBonus(jsonObject.getInteger("bonus"));
+            stu.setBonus(jsonObject.getBigDecimal("bonus"));
             studentService._updateStudentById(stu);
         }
         RenderKit.renderSuccess(this);
